mcp-server-github/requirements.txt
mcp>=1.1.2
PyGithub>=2.1.1
pydantic>=2.5.0